key-server/requirements.txt
fastapi
uvicorn
pydantic
duckdb
python-dotenv
web3
python-multipart
gunicorn